+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 ### Added
-- Added displayProblem to print problem to stdout
+- Added categorical data example
+- Added printProblem to print problem to stdout
 - Added stage checks to presolve, freereoptsolve, freetransform
 - Added primal_dual_evolution recipe and a plot recipe
 ### Fixed
-- Fixed default name for indicator constraints
 ### Changed
+- GitHub actions using Mac now use precompiled SCIP from latest release
 ### Removed
 
 ## 5.2.1 - 2024.10.29